# docs/specification/openapi.yaml
openapi: 3.1.0
info:
  title: TestModeller API
  version: 0.1.0
  description: >
    Contract between the Angular UI and the Rust backend. Changes require explicit approval (see AGENTS.md).
    Hierarchy: Project > Component > Feature > Model. A Feature has a scenario description and one or more Models
    (state machines). Test cases belong to a Feature and are assigned to Model states or transitions (steps).
servers:
  - url: /api/v1
...
paths:
...
  /features/{featureId}/models:
    parameters: [ { $ref: '#/components/parameters/FeatureId' } ]
    get:
      tags: [Models]
      summary: List models of a feature (summaries)
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ModelSummary' } } } } }
    post:
      tags: [Models]
      summary: Create model (optionally with states and transitions)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ModelInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Model' } } } }
  /models/{modelId}:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [Models]
      summary: Get full model with states and transitions
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Model' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [Models]
      summary: Replace whole model graph atomically (editor save)
      parameters: [ { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ModelInput' } } } }
      responses:
        '200': { description: Saved, content: { application/json: { schema: { $ref: '#/components/schemas/Model' } } } }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Models]
      summary: Update model metadata only (name, description, status)
      parameters: [ { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ModelMetaInput' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/ModelSummary' } } } }
        '412': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Models]
      summary: Delete model (assignments are removed, test cases kept)
      responses:
        '204': { description: Deleted }
  /models/{modelId}/duplicate:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [Models]
      summary: Duplicate model, optionally into another feature
      requestBody:
        content: { application/json: { schema: { type: object, properties: { name: { type: string }, featureId: { $ref: '#/components/schemas/Id' } } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Model' } } } }
  /models/{modelId}/validate:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [Models]
      summary: Validate stored model, or a draft graph if a body is given
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/ModelInput' } } }
      responses:
        '200': { description: Issues (empty if valid), content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } } } } }
  /models/{modelId}/versions:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [Models]
      summary: Version history
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ModelVersion' } } } } }
  /models/{modelId}/versions/{version}/restore:
    parameters:
      - { $ref: '#/components/parameters/ModelId' }
      - { name: version, in: path, required: true, schema: { type: integer, minimum: 1 } }
    post:
      tags: [Models]
      summary: Restore a previous version as a new version
      responses:
        '200': { description: Restored, content: { application/json: { schema: { $ref: '#/components/schemas/Model' } } } }
  /models/{modelId}/states:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [States]
      summary: List states
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/State' } } } } }
    post:
      tags: [States]
      summary: Add state
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StateInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/State' } } } }
  /states/{stateId}:
    parameters: [ { $ref: '#/components/parameters/StateId' } ]
    get:
      tags: [States]
      summary: Get state
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/State' } } } }
    patch:
      tags: [States]
      summary: Update state
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StateInput' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/State' } } } }
    delete:
      tags: [States]
      summary: Delete state and its incident transitions
      responses:
        '204': { description: Deleted }
  /models/{modelId}/transitions:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [Transitions]
      summary: List transitions (model steps)
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Transition' } } } } }
    post:
      tags: [Transitions]
      summary: Add transition
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TransitionInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Transition' } } } }
  /transitions/{transitionId}:
    parameters: [ { $ref: '#/components/parameters/TransitionId' } ]
    get:
      tags: [Transitions]
      summary: Get transition
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Transition' } } } }
    patch:
      tags: [Transitions]
      summary: Update transition
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TransitionInput' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Transition' } } } }
    delete:
      tags: [Transitions]
      summary: Delete transition
      responses:
        '204': { description: Deleted }
...
components:
  parameters:
...
    FeatureId: { name: featureId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    ModelId: { name: modelId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    StateId: { name: stateId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    TransitionId: { name: transitionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Entity `version` for optimistic concurrency
      schema: { type: integer, minimum: 1 }
...
  responses:
    Problem:
      description: Error (RFC 7807)
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
  schemas:
    Id: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
...
    Audit:
      type: object
      required: [id, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        version: { type: integer, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    ModelStatus: { type: string, enum: [draft, ready] }
    ModelMetaInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string }
        status: { $ref: '#/components/schemas/ModelStatus' }
    StateKind: { type: string, enum: [initial, normal, final] }
    Position:
      type: object
      required: [x, y]
      properties: { x: { type: number }, y: { type: number } }
    StateInput:
      type: object
      required: [name, kind]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name: { type: string, minLength: 1 }
        description: { type: string }
        kind: { $ref: '#/components/schemas/StateKind' }
        position: { $ref: '#/components/schemas/Position' }
    State:
      allOf:
        - { $ref: '#/components/schemas/StateInput' }
        - type: object
          required: [id, modelId, testCaseCount]
          properties:
            id: { $ref: '#/components/schemas/Id' }
            modelId: { $ref: '#/components/schemas/Id' }
            testCaseCount: { type: integer, readOnly: true }
    TransitionInput:
      type: object
      required: [from, to, event]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        from: { $ref: '#/components/schemas/Id' }
        to: { $ref: '#/components/schemas/Id' }
        event: { type: string, minLength: 1 }
        guard: { type: string }
        action: { type: string }
        expected: { type: string }
    Transition:
      allOf:
        - { $ref: '#/components/schemas/TransitionInput' }
        - type: object
          required: [id, modelId, testCaseCount]
          properties:
            id: { $ref: '#/components/schemas/Id' }
            modelId: { $ref: '#/components/schemas/Id' }
            testCaseCount: { type: integer, readOnly: true }
    Variable:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string, enum: [boolean, integer, string] }
        initial: {}
    ModelInput:
      allOf:
        - { $ref: '#/components/schemas/ModelMetaInput' }
        - type: object
          required: [name]
          properties:
            variables: { type: array, items: { $ref: '#/components/schemas/Variable' } }
            states: { type: array, items: { $ref: '#/components/schemas/StateInput' } }
            transitions: { type: array, items: { $ref: '#/components/schemas/TransitionInput' } }
    ModelSummary:
      allOf:
        - { $ref: '#/components/schemas/Audit' }
        - { $ref: '#/components/schemas/ModelMetaInput' }
        - type: object
          required: [featureId, stateCount, transitionCount, testCaseCount]
          properties:
            featureId: { $ref: '#/components/schemas/Id' }
            stateCount: { type: integer }
            transitionCount: { type: integer }
            testCaseCount: { type: integer }
    Model:
      allOf:
        - { $ref: '#/components/schemas/ModelSummary' }
        - type: object
          properties:
            variables: { type: array, items: { $ref: '#/components/schemas/Variable' } }
            states: { type: array, items: { $ref: '#/components/schemas/State' } }
            transitions: { type: array, items: { $ref: '#/components/schemas/Transition' } }
    ModelVersion:
      type: object
      required: [version, createdAt]
      properties: { version: { type: integer }, createdAt: { $ref: '#/components/schemas/Timestamp' }, summary: { type: string } }
    ValidationIssue:
      type: object
      required: [severity, code, message]
      properties:
        severity: { type: string, enum: [error, warning] }
        code: { type: string, examples: [NO_INITIAL_STATE, MULTIPLE_INITIAL_STATES, UNREACHABLE_STATE, DEAD_END, BAD_GUARD] }
        message: { type: string }
        stateId: { $ref: '#/components/schemas/Id' }
        transitionId: { $ref: '#/components/schemas/Id' }
...